Simplify the AbstractNetworkTest tree (undirected branch): Delete StandardSimpleUndirectedNetworkTest
This is similar to an earlier refector of the AbstractGraphTest tree.

Note: This is a pure refactor: No tests were added or removed.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=300055427
